Issue #434 fix bug

+++ tutor/app/assets/javascripts/resources.js.coffee
@@ -2,12 +2,13 @@ $(document).ready ->
 	$("#resources").on "cocoon:before-remove", (e, resource) ->
 		$(this).data "remove-timeout", 1000
 		x = resource[0].children[1].id.replace("link","id")
-		id = $("#"+x)[0].value
-		$.ajax
-			url: id
-			type: "DELETE"
-			success: (result) ->
-				document.getElementById(x).remove()
+		if $("#"+x)[0]
+			id = $("#"+x)[0].value
+			$.ajax
+				url: id
+				type: "DELETE"
+				success: (result) ->
+					document.getElementById(x).remove()
 		resource.fadeOut "slow"
 
 	$("#resources").on "cocoon:after-remove", (e, resource) ->
